Incorporate the comments
