--- page_classes/base.page.js
@@ -1,7 +1,0 @@
-class BasePage {
-    open(path){
-        browser.url(path);
-    }
-}
-
-export default BasePage;
