+++ pages/home/selectors.js
@@ -1,0 +1,8 @@
+class Selectors {
+
+    get flash(){ return  $('#flash'); }
+    get logout(){ return $('=Logout'); }
+
+}
+
+export default new Selectors();
