+++ tests/login_tests/login.test.js
@@ -1,4 +1,4 @@
-import loginPage from '../../page_classes/login.page';
+import loginPage from '../../pages/login/login.page';
 
 describe('Tests related to Login page', () => {
     beforeEach('Navigate to Login page', () => {
@@ -7,10 +7,12 @@ describe('Tests related to Login page', () => {
 
     it('Login should failed with invalid credentials.', () => {
         loginPage.login("qwerr","dsdsds");
+        loginPage.checkElementalSelenium();
     });
 
     it('Login should successful with valid credentials.', () => {
         let homePage = loginPage.login("tomsmith", "SuperSecretPassword!");
+        homePage.checkElementalSelenium();
         homePage.signout();
     });
